Add a version tag jira filter changelog generator.
This creates a URL with a jql query that will look for jira versions that
match git tag names in the commit range. It currently assumes that all
tags found will have a corresponding jira version. The jql will error if
this is not the case. The URL will also filter to specific projects, to
avoid version number clashes with other projects.